Migrate to TypeScript, new Slack SDK, node fetch and update other dependencies
