+++ package.json
@@ -5,23 +5,24 @@
   "main": "index.js",
   "repository": "git@github.com:blankoslo/floq-timebot.git",
   "author": "Erlend Hamberg",
-  "license": "Unlicenced",
+  "license": "UNLICENCED",
   "scripts": {
-    "lint": "eslint .",
-    "flow": "flow check"
+    "ts": "tsc --noEmit",
+    "build": "tsup index.ts",
+    "lint": "eslint ."
   },
   "dependencies": {
-    "@slack/client": "^3.8.1",
-    "date-fns": "^1.22.0",
-    "jsonwebtoken": "^7.2.1",
-    "request": "^2.79.0",
-    "request-promise-native": "^1.0.3"
+    "@slack/web-api": "^6.8.1",
+    "@types/jsonwebtoken": "^9.0.1",
+    "date-fns": "^2.29.3",
+    "jsonwebtoken": "^9.0.0",
+    "node-fetch": "2"
   },
   "devDependencies": {
-    "babel-eslint": "^7.1.1",
-    "eslint": "^3.12.2",
-    "eslint-config-walmart": "^1.1.0",
-    "eslint-plugin-filenames": "^1.1.0",
-    "flow-bin": "^0.37.4"
+    "eslint": "^8.36.0",
+    "eslint-config-walmart": "^2.2.1",
+    "eslint-plugin-filenames": "^1.3.2",
+    "tsup": "^6.7.0",
+    "typescript": "^5.0.2"
   }
 }
